Update README.md
Changes to the Hello World file as part of the Johns Hopkins', "The Data Scientist's Toolbox" course at USDA Coursera.  This is part of the preparation to become a better coder specifically in R programming.  For the USDA purposes it is so I can pick up the RNA Sequencing pipeline at least for now.